+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 Version 3.1.4-dev
 -----------------
 
-Nothing yet.
+### Fixed
+
+* Fixed pretty printing of `-(-$x)` and `+(+$x)`. (#459)
 
 Version 3.1.3 (2017-12-26)
 --------------------------